requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
Django==5.1.3
django-allauth==65.2.0
django-crispy-forms==2.3
django-tailwind==3.8.0
psycopg2-binary==2.9.10
sqlparse==0.5.2
